Split ContextMenu out to TS file
